+++ CHANGELOG.rst
@@ -27,7 +27,7 @@ Changed
 -------
 
 - **General**
-    - Upgrade to django-sodar-core v1.2.0 (#2187)
+    - Upgrade to django-sodar-core v1.2.1 (#2187)
     - Upgrade general Python depedencies (#2187)
     - Refactor plugin helpers to use ``PluginAPI`` (#2187)
     - Move black configuration to ``pyproject.toml`` (#2187)